requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
marshmallow==3.15.0

numpy~=1.22.3
PyYAML~=5.4.1
setuptools~=57.0.0
tqdm~=4.64.0
pillow~=9.1.0

scikit-learn~=1.0.2
torch~=1.11.0
mlflow~=1.25.1
xgboost~=1.6.1
boto3~=1.22.9
pytest